Changed unsigned to size_t

+++ Util/fft/FFTW3_Transform.h
@@ -20,21 +20,21 @@ namespace FTransform {
   public:
 
     //! Constructor
-    FFTW3_Plan (unsigned nfft, const std::string& call);
+    FFTW3_Plan (size_t nfft, const std::string& call);
 
     //! Destructor
     ~FFTW3_Plan ();
 
-    static int fcc1d (unsigned nfft, float* dest, const float* src);
-    static int bcc1d (unsigned nfft, float* dest, const float* src);
-    static int frc1d (unsigned nfft, float* dest, const float* src);
-    static int bcr1d (unsigned nfft, float* dest, const float* src);
+    static int fcc1d (size_t nfft, float* dest, const float* src);
+    static int bcc1d (size_t nfft, float* dest, const float* src);
+    static int frc1d (size_t nfft, float* dest, const float* src);
+    static int bcr1d (size_t nfft, float* dest, const float* src);
 
     //! Agent class
     class Agent : public PlanAgent<FFTW3_Plan> {
     public:
       Agent () : PlanAgent<FFTW3_Plan> ("FFTW3", nfft) { }
-      FFTW3_Plan* new_plan (unsigned nfft, const std::string& call);
+      FFTW3_Plan* new_plan (size_t nfft, const std::string& call);
     };
 
   protected:
@@ -49,20 +49,20 @@ namespace FTransform {
   public:
 
     //! Constructor
-    FFTW3_Plan2 (unsigned nx, unsigned ny, const std::string& call);
+    FFTW3_Plan2 (size_t nx, size_t ny, const std::string& call);
 
     //! Destructor
     ~FFTW3_Plan2 ();
 
-    static void fcc2d (unsigned nx, unsigned ny, float* , const float* );
-    static void bcc2d (unsigned nx, unsigned ny, float* , const float* );
+    static void fcc2d (size_t nx, size_t ny, float* , const float* );
+    static void bcc2d (size_t nx, size_t ny, float* , const float* );
 
     class Agent : public PlanAgent2<FFTW3_Plan2> {
     public:
       Agent () : PlanAgent2<FFTW3_Plan2> ("FFTW3", nfft) { }
 
       FFTW3_Plan2*
-      new_plan (unsigned nx, unsigned ny, const std::string& call);
+      new_plan (size_t nx, size_t ny, const std::string& call);
     };
 
   protected:
