+++ Util/fft/MKL_Transform.h
@@ -18,22 +18,22 @@ namespace FTransform {
   public:
 
     //! Constructor
-    MKL_Plan (unsigned nfft, const std::string& _fft_call);
+    MKL_Plan (size_t nfft, const std::string& _fft_call);
 
     //! Destructor
     ~MKL_Plan ();
 
-    static int fcc1d (unsigned nfft, float* dest, const float* src);
-    static int bcc1d (unsigned nfft, float* dest, const float* src);
-    static int frc1d (unsigned nfft, float* dest, const float* src);
-    static int bcr1d (unsigned nfft, float* dest, const float* src);
+    static int fcc1d (size_t nfft, float* dest, const float* src);
+    static int bcc1d (size_t nfft, float* dest, const float* src);
+    static int frc1d (size_t nfft, float* dest, const float* src);
+    static int bcr1d (size_t nfft, float* dest, const float* src);
 
     //! Agent class
     class Agent : public PlanAgent<MKL_Plan> {
     public:
       Agent ();
       virtual ~Agent();
-      MKL_Plan* new_plan (unsigned nfft, const std::string& call);
+      MKL_Plan* new_plan (size_t nfft, const std::string& call);
     };
 
   protected:
